Removed LtiProviderFinder and LtiProviderResource (overkill).
