+++ models/classes/LtiProvider/RdfLtiProviderRepository.php
@@ -20,74 +20,120 @@
 namespace oat\taoLti\models\classes\LtiProvider;
 
 use oat\generis\model\OntologyRdfs;
-use oat\oatbox\service\ConfigurableService;
-use oat\oatbox\service\exception\InvalidServiceManagerException;
+use oat\tao\model\OntologyClassService;
+use common_exception_Error as ErrorException;
+use common_exception_InvalidArgumentType as InvalidArgumentTypeException;
+use core_kernel_classes_Resource as RdfResource;
+use oat\generis\model\kernel\persistence\smoothsql\search\ComplexSearchService;
+use oat\tao\model\oauth\DataStore;
 
 /**
  * Service methods to manage the LTI provider business objects.
  */
-class RdfLtiProviderRepository extends ConfigurableService implements LtiProviderRepositoryInterface
+class RdfLtiProviderRepository extends OntologyClassService implements LtiProviderRepositoryInterface
 {
+    const CLASS_URI = 'http://www.tao.lu/Ontologies/TAOLTI.rdf#LTIProvider';
+
+    public function getRootClass()
+    {
+        return $this->getClass(self::CLASS_URI);
+    }
+
     public function count()
     {
-        try {
-            return $this->getFinderService()->getResourcesCount();
-        } catch (InvalidServiceManagerException $exception) {
-            $this->logError('Unable to retrieve providers: ' . $exception->getMessage());
-
-            return 0;
-        }
+        return $this->queryResources([], 'count', 0);
     }
 
     public function findAll()
     {
-        return $this->findBy();
+        return $this->getProviders();
     }
 
     public function searchByLabel($queryString)
     {
-        return $this->findBy([OntologyRdfs::RDFS_LABEL => $queryString]);
+        return $this->getProviders([OntologyRdfs::RDFS_LABEL => $queryString]);
     }
 
     /**
-     * Retrieves LTI Providers corresponding to the given criteria.
+     * Retrieves providers from rdf store corresponding to the given criteria.
      *
      * @param array $criteria
      *
-     * @return array
+     * @return array|LtiProvider[]
      */
-    public function findBy(array $criteria = [])
+    private function getProviders(array $criteria = [])
     {
+        $resources = $this->queryResources($criteria, 'search', []);
+        $ltiProviders = [];
+
         try {
-            $resources = $this->getFinderService()->getResources($criteria);
-        } catch (InvalidServiceManagerException $exception) {
-            $this->logError('Unable to retrieve providers: ' . $exception->getMessage());
+            foreach ($resources as $resource) {
+                $ltiProviders[] = $this->getLtiProviderFromResource($resource);
+            }
+        } catch (InvalidArgumentTypeException $exception) {
+            $this->logError('Unable to retrieve provider properties: ' . $exception->getMessage());
 
             return [];
         }
 
-        return array_map(
-            function (LtiProviderResource $resource) {
-                return new LtiProvider(
-                    $resource->getUri(),
-                    $resource->getLabel(),
-                    $resource->getKey(),
-                    $resource->getSecret(),
-                    $resource->getCallbackUrl()
-                );
-            },
-            $resources
-        );
+        return $ltiProviders;
     }
 
     /**
-     * Returns search service.
+     * Retrieves resources from rdf store corresponding to the given criteria,
+     * hydrate the result with the $hydration method
+     * or returns $default on failure
      *
-     * @return ConfigurableService|RdfLtiProviderFinder
-     * @throws InvalidServiceManagerException
+     * @param array     $criteria
+     * @param string    $hydration Hydration method ("search" for actual results, "count" for counting results)
+     * @param array|int $default   default value to return on failure
+     *
+     * @return mixed
      */
-    private function getFinderService()
+    private function queryResources(array $criteria, $hydration, $default)
     {
-        return $this->getServiceManager()->get(RdfLtiProviderFinder::class);
+        try {
+            /** @var ComplexSearchService $searchService */
+            $searchService = $this->getServiceLocator()->get(ComplexSearchService::SERVICE_ID);
+            $queryBuilder = $searchService->query();
+            $query = $searchService->searchType($queryBuilder, self::CLASS_URI, true);
+            if (count($criteria)) {
+                foreach ($criteria as $property => $value) {
+                    $query->add($property)->contains($value);
+                }
+            }
+            $queryBuilder->setCriteria($query);
+
+            return $searchService->getGateway()->$hydration($queryBuilder);
+        } catch (ErrorException $e) {
+            $this->logError('Unable to retrieve providers: ' . $e->getMessage());
+
+            return $default;
+        }
+    }
+
+    /**
+     * Completes the RdfResource by adding the key, secret and callbackUrl properties.
+     *
+     * @param RdfResource $resource Base RdfResource providing Uri and Label.
+     *
+     * @return LtiProvider
+     * @throws InvalidArgumentTypeException
+     */
+    private function getLtiProviderFromResource(RdfResource $resource)
+    {
+        $propertiesValues = $resource->getPropertiesValues([
+            DataStore::PROPERTY_OAUTH_KEY,
+            DataStore::PROPERTY_OAUTH_SECRET,
+            DataStore::PROPERTY_OAUTH_CALLBACK,
+        ]);
+
+        return new LtiProvider(
+            $resource->getUri(),
+            $resource->getLabel(),
+            (string)reset($propertiesValues[DataStore::PROPERTY_OAUTH_KEY]),
+            (string)reset($propertiesValues[DataStore::PROPERTY_OAUTH_SECRET]),
+            (string)reset($propertiesValues[DataStore::PROPERTY_OAUTH_CALLBACK])
+        );
     }
 }
